:label: 修改缺陷修复时长数据类型为 int

+++ quick-click-server/src/main/java/io/github/xiaomisum/quickclick/dal/dataobject/quality/Bug.java
@@ -31,7 +31,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import xyz.migoo.framework.mybatis.core.dataobject.BaseDO;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @EqualsAndHashCode(callSuper = true)
@@ -132,7 +131,7 @@ public class Bug extends BaseDO<String> {
     /**
      * 修复时长
      */
-    private BigDecimal fixDuration;
+    private Integer fixDuration;
 
     /**
      * 关闭人
